refactor: receipt-check,write,plan URL 수정
