+++ src/pages/index.ts
@@ -1,12 +1,12 @@
 export { default as Login } from "./login";
 export { default as Home } from "./home";
 export { default as Mypage } from "./mypage";
-export { default as Schedule } from "./schedule";
+export { default as Plan } from "./plan";
 export { default as Place } from "./place";
-export { default as KakaoCallback } from './auth/kakao-callback';
-export { default as OnboardingUser} from './onboarding/user';
-export { default as OnboardingPet} from './onboarding/pet';
-export { default as OnboardingCompletion} from './onboarding/completion';
-export { default as MyReview} from './mypage/my-review';
-export { default as NotFound} from './not-found';
-export { default as MyBookMark} from './mypage/bookmark';
+export { default as KakaoCallback } from "./auth/kakao-callback";
+export { default as OnboardingUser } from "./onboarding/user";
+export { default as OnboardingPet } from "./onboarding/pet";
+export { default as OnboardingCompletion } from "./onboarding/completion";
+export { default as MyReview } from "./mypage/my-review";
+export { default as NotFound } from "./not-found";
+export { default as MyBookMark } from "./mypage/bookmark";
