+++ src/pages/plan/index.styles.ts
